# swagger.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "Authentication API",
		"description": "A sample API to demonstrate Swagger/OpenAPI specification",
		"version": "1.0.0",
		"contact": {
			"email": "contact@example.com"
		}
	},
	"servers": [
		{
			"url": "http://localhost:3000",
			"description": "Local development server"
		}
	],
...
	"components": {
		"securitySchemes": {
			"BearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"description": "Enter your JWT token in the format: Bearer <token>"
			}
		},
		"schemas": {
			"SignupRequest": {
				"type": "object",
				"properties": {
					"email": {
						"type": "string",
						"format": "email",
						"description": "User's email address"
					},
					"firstName": {
						"type": "string",
						"description": "User's first name"
					},
					"lastName": {
						"type": "string",
						"description": "User's last name"
					},
					"phoneNumberCode": {
						"type": "string",
						"description": "Country code for phone number"
					},
					"phoneNumber": {
						"type": "string",
						"description": "User's phone number"
					},
					"password": {
						"type": "string",
						"format": "password",
						"description": "User's password"
					}
				},
				"required": ["email", "firstName", "lastName", "password"]
			},
			"SignupResponse": {
				"type": "object",
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"type": "object",
						"properties": {
							"user": {
								"type": "object",
								"properties": {
									"id": {
										"type": "integer"
									},
									"email": {
										"type": "string",
										"format": "email"
									},
									"firstName": {
										"type": "string"
									},
									"lastName": {
										"type": "string"
									},
									"phoneNumberCode": {
										"type": "string"
									},
									"phoneNumber": {
										"type": "string"
									},
									"isEmailVerified": {
										"type": "boolean"
									},
									"isPhoneVerified": {
										"type": "boolean"
									},
									"createdAt": {
										"type": "string",
										"format": "date-time"
									}
								}
							},
							"token": {
								"type": "string"
							}
						}
					}
				}
			},
...
		}
	},
	"paths": {
		"/api/v1/user/signup": {
			"post": {
				"tags": ["Auth"],
				"summary": "User signup",
				"description": "Register a new user account",
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SignupRequest"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "User registered successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SignupResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					}
				}
			}
		},
		"/api/v1/user/login": {
			"post": {
				"tags": ["Auth"],
				"summary": "User login",
				"description": "Authenticate a user and get access token",
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SigninRequest"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Login successful",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SigninResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					}
				}
			}
		},
		"/api/v1/user/social-login": {
			"post": {
				"tags": ["Auth"],
				"summary": "Social login",
				"description": "Login using a social provider (Google, Apple)",
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SocialLoginRequest"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Login successful",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SigninResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					}
				}
			}
		},
		"/api/v1/user/profile": {
			"get": {
				"tags": ["User"],
				"summary": "Get user profile",
				"description": "Retrieves the profile information of the logged-in user",
				"operationId": "getProfile",
				"security": [{ "BearerAuth": [] }],
				"parameters": [
					{
						"in": "header",
						"name": "Authorization",
						"schema": { "type": "string" },
						"required": true,
						"description": "JWT token (Bearer <token>)",
						"example": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
					}
				],
				"responses": {
					"200": {
						"description": "Profile fetched successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProfileResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized access",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					}
				}
			},
			"put": {
				"tags": ["User"],
				"summary": "Update user profile",
				"description": "Updates the profile information of the logged-in user",
				"operationId": "updateProfile",
				"security": [{ "BearerAuth": [] }],
				"parameters": [
					{
						"in": "header",
						"name": "Authorization",
						"schema": { "type": "string" },
						"required": true,
						"description": "JWT token (Bearer <token>)",
						"example": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateProfileRequest"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Profile updated successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProfileResponse"
								}
							}
						}
					},
					"400": {
						"description": "Invalid input",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized access",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					}
				}
			}
		},
		"/api/v1/user/password": {
			"put": {
				"tags": ["User"],
				"summary": "Change user password",
				"description": "Change the password of logged-in user",
				"operationId": "changePassword",
				"security": [{ "BearerAuth": [] }],
				"parameters": [
					{
						"in": "header",
						"name": "Authorization",
						"schema": { "type": "string" },
						"required": true,
						"description": "JWT token (Bearer <token>)",
						"example": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChangePasswordRequest"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Password changed successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					},
					"400": {
						"description": "Invalid password",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized access",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					}
				}
			}
		},
		"/api/v1/user/language": {
			"put": {
				"tags": ["User"],
				"summary": "Change user language preference",
				"description": "Change the language preference for the logged-in user",
				"operationId": "changeLanguage",
				"security": [{ "BearerAuth": [] }],
				"parameters": [
					{
						"in": "header",
						"name": "Authorization",
						"schema": { "type": "string" },
						"required": true,
						"description": "JWT token (Bearer <token>)",
						"example": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChangeLanguageRequest"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Language preference updated successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					},
					"400": {
						"description": "Invalid language code",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized access",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					}
				}
			}
		},
		"/api/v1/user/social-link": {
			"post": {
				"tags": ["User"],
				"summary": "Link social account",
				"description": "Link a social account to the user's profile",
				"operationId": "linkSocialAccount",
				"security": [{ "BearerAuth": [] }],
				"parameters": [
					{
						"in": "header",
						"name": "Authorization",
						"schema": { "type": "string" },
						"required": true,
						"description": "JWT token (Bearer <token>)",
						"example": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LinkSocialAccountRequest"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Social account linked successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					},
					"400": {
						"description": "Invalid input or token",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized access",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					}
				}
			}
		},
		"/api/v1/user/notification-settings": {
			"get": {
				"tags": ["User"],
				"summary": "Get notification settings",
				"description": "Retrieve user's notification preferences",
				"operationId": "getNotificationSettings",
				"security": [{ "BearerAuth": [] }],
				"parameters": [
					{
						"in": "header",
						"name": "Authorization",
						"schema": { "type": "string" },
						"required": true,
						"description": "JWT token (Bearer <token>)",
						"example": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
					}
				],
				"responses": {
					"200": {
						"description": "Notification settings retrieved successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotificationSettings"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized access",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					}
				}
			},
			"patch": {
				"tags": ["User"],
				"summary": "Update notification settings",
				"description": "Update user's notification preferences",
				"operationId": "updateNotificationSettings",
				"security": [{ "BearerAuth": [] }],
				"parameters": [
					{
						"in": "header",
						"name": "Authorization",
						"schema": { "type": "string" },
						"required": true,
						"description": "JWT token (Bearer <token>)",
						"example": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/NotificationSettings"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Notification settings updated successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotificationSettings"
								}
							}
						}
					},
					"400": {
						"description": "Invalid input",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized access",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					}
				}
			}
		},
...
	}
}